// src/petals/types.ts
export type Petal = {
    x: number;
    y: number;
    size: number;
    depth: number;
    rotation: number;
    rotationSpeed: number;
    flipAngle: number;
    flipSpeed: number;
    swingAmplitude: number;
    swingFrequency: number;
    swingOffset: number;
    fallSpeed: number;
    colorIndex: number;
};